refactor(misc): dcat normalize/validate/context/convert to ts (+ spec import paths)
