+++ dev/type-errors-baseline.txt
@@ -1,1 +1,1 @@
-1360
+1354
